+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename `-g` and `--grep` flags to `-s` and `--select`.
 - Group output logs using regular expressions via the `-g` or `--group` flag.
 - Implicitly filter based on given fields and group.
+- Add golden file tests to automatically verify certain features.